Updates links, adds missing topics, and summary
This updates aims the following objectives

- Adds some missing topics that were not present in the markdown.

- Some topics were not in the same order as we can find from the DCA blueprint.
Because this can cause confusion in the future when maintaining this markdown,
I re-orderd them.

- Some links were switched either because they were being redirect to a newer
documentation link at the Docker Official Documentation or because I found an
official explanation from the Docker Official Documentation. The objective
is to avoid using a possibly outdated information.

- Added a quick facts section regarding to the exam, so we don't need to
navigate away from this Markdown.